+++ sources/www/widgets/state-display/viewModel.js
@@ -31,9 +31,9 @@ function stateDisplayViewModel() {
        var defferedConfigurationChangedFinished = new $.Deferred();
         
         //we register keyword new acquisition
-        self.widgetApi.registerKeywordForNewAcquisitions(self.widget.configuration.device.keywordId);
-		  self.widgetApi.getLastValue(self.widget.configuration.device.keywordId); 		
-        self.widgetApi.configureBatteryIcon(self.widget.configuration.device.deviceId);
+        self.widgetApi.registerKeywordForNewAcquisitions(parseInt(self.widget.configuration.device.keywordId));
+		  self.widgetApi.getLastValue(parseInt(self.widget.configuration.device.keywordId));
+        self.widgetApi.configureBatteryIcon(parseInt(self.widget.configuration.device.deviceId));
         self.widgetApi.registerAdditionalInformation(["pluginId", "typeInfo"]);
     }
 
@@ -45,7 +45,7 @@ function stateDisplayViewModel() {
     this.onNewAcquisition = function (keywordId, data) {
         var self = this;
         
-        if (keywordId !== self.widget.configuration.device.keywordId)
+        if (keywordId !== parseInt(self.widget.configuration.device.keywordId))
            return;
         
         if (!isNullOrUndefinedOrEmpty(data.typeInfo))