+++ sources/www/widgets/wind/viewModel.js
@@ -128,12 +128,12 @@ widgetViewModelCtor = function windViewModel() {
       var self = this;
 
       if (parseBool(self.widget.configuration.speedDisplay.checkbox)) {
-         if (keywordId === self.widget.configuration.speedDisplay.content.windSpeed.keywordId) {
+         if (keywordId === parseInt(self.widget.configuration.speedDisplay.content.windSpeed.keywordId)) {
             self.windspeed(self.formatSpeed(data.value));
          }
       }
 
-      if (keywordId === self.widget.configuration.windDirection.keywordId) {
+      if (keywordId === parseInt(self.widget.configuration.windDirection.keywordId)) {
          self.direction = self.mod(parseFloat(data.value) + self.correction, 360);
          self.displayNeedle(self.direction);
       }
@@ -152,12 +152,12 @@ widgetViewModelCtor = function windViewModel() {
       if ((isNullOrUndefined(self.widget)) || (isNullOrUndefinedOrEmpty(self.widget.configuration)))
          return;
 
-      self.widgetApi.registerKeywordForNewAcquisitions([self.widget.configuration.windDirection.keywordId]);
-      self.widgetApi.getLastValue([self.widget.configuration.windDirection.keywordId]);
+      self.widgetApi.registerKeywordForNewAcquisitions([parseInt(self.widget.configuration.windDirection.keywordId)]);
+      self.widgetApi.getLastValue([parseInt(self.widget.configuration.windDirection.keywordId)]);
 
       if (parseBool(self.widget.configuration.speedDisplay.checkbox)) {
-         self.widgetApi.registerKeywordForNewAcquisitions([self.widget.configuration.speedDisplay.content.windSpeed.keywordId]);
-         self.widgetApi.getLastValue([self.widget.configuration.speedDisplay.content.windSpeed.keywordId]);
+         self.widgetApi.registerKeywordForNewAcquisitions([parseInt(self.widget.configuration.speedDisplay.content.windSpeed.keywordId)]);
+         self.widgetApi.getLastValue([parseInt(self.widget.configuration.speedDisplay.content.windSpeed.keywordId)]);
          self.widgetApi.find(".windspeed").css("visibility", "visible");
       } else
          self.widgetApi.find(".windspeed").css("visibility", "hidden");