+++ tests/unit/sources/plugins/dev-FakePlugin/TestFakePluginConfiguration.cpp
@@ -9,23 +9,27 @@ BOOST_AUTO_TEST_SUITE(TestFakePluginConfiguration)
 
 
 const shared::CDataContainer conf("{"
-   "\"EnumParameter\": \"EnumValue1\""
-   "}");
+	"\"EnumParameter\": \"EnumValue1\""
+	"}");
 
 BOOST_AUTO_TEST_CASE(ReadFakePluginDefaultConfiguration)
 {
-   // Need to copy the package.json file from fakePlugin, needed by initializeWith method to get default configuration
-	if (boost::filesystem::exists("../../../../sources/plugins/dev-FakePlugin/package.in.json"))
-		boost::filesystem::copy_file("../../../../sources/plugins/dev-FakePlugin/package.in.json", "package.json", boost::filesystem::copy_option::overwrite_if_exists);
-	else if (boost::filesystem::exists("../../../sources/plugins/dev-FakePlugin/package.in.json"))
+	// Need to copy the package.json file from fakePlugin, needed by initializeWith method to get default configuration
+	if (boost::filesystem::exists("../../../sources/plugins/dev-FakePlugin/package.in.json")) {
 		boost::filesystem::copy_file("../../../sources/plugins/dev-FakePlugin/package.in.json", "package.json", boost::filesystem::copy_option::overwrite_if_exists);
+	}
+	else if (boost::filesystem::exists("../../../../sources/plugins/dev-FakePlugin/package.in.json")) {
+		boost::filesystem::copy_file("../../../../sources/plugins/dev-FakePlugin/package.in.json", "package.json", boost::filesystem::copy_option::overwrite_if_exists);
+	}
 	else
-		BOOST_ERROR("Fail to copy fakePlugin package.in.json to package.json");
+	{
+		BOOST_ERROR("dev-FakePlugin/package.in.json not found");
+	}
 
-   CFakePluginConfiguration cfg;
-   cfg.initializeWith(conf);
+	CFakePluginConfiguration cfg;
+	cfg.initializeWith(conf);
 
-   BOOST_CHECK_EQUAL(cfg.getEnumParameter(), static_cast<EEnumType>(kEnumValue1));
+	BOOST_CHECK_EQUAL(cfg.getEnumParameter(), static_cast<EEnumType>(kEnumValue1));
 }
 
 BOOST_AUTO_TEST_SUITE_END()