+++ tests/unit/sources/shared/TestDataContainer.cpp
@@ -225,6 +225,36 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
       BOOST_CHECK_EQUAL_COLLECTIONS(allconditions.begin(), allconditions.end(), getAllCond.begin(), getAllCond.end()) ;
    }
 
+   BOOST_AUTO_TEST_CASE(ContainerToVectorOfContainers)
+   {
+      const shared::CDataContainer inIsNotAnArray("\
+      {\
+         \"changelogUrl\": \"http://www.yadoms.com/downloads/update/widgets/moon/changelog.md\",\
+            \"0\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0\"\
+         },\
+         \"1\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.2\"\
+         },\
+         \"2\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0-rc.1\"\
+         },\
+         \"3\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.1\"\
+         }\
+      }");
+      BOOST_CHECK_THROW(const auto out = inIsNotAnArray.get<std::vector<shared::CDataContainer>>(), shared::exception::COutOfRange);
+   }
+
+
    boost::shared_ptr<shared::serialization::IDataSerializable> maketest(const unsigned int testcount)
    {
 	   shared::CDataContainer result;
@@ -802,13 +832,11 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
    }
 
 
-   // TOFIX : This test is disabled because the default it revealed is not so serious and
-   // it makes fail to full campaign. We have for the moment no good solution to fix it.
-   // A good solution would probably to change our JSON parser (from Boost) by a real JSON parser.
-   //BOOST_AUTO_TEST_CASE(SimpleConstruction)
-   //{
-   //   BOOST_CHECK_THROW(shared::CDataContainer dc("1"), std::exception) ;
-   //}
+   BOOST_AUTO_TEST_CASE(SimpleConstruction)
+   {
+      shared::CDataContainer dc("1");
+      BOOST_CHECK_EQUAL(dc.serialize(), "1");
+   }
 
 
    BOOST_AUTO_TEST_CASE(Merge)
@@ -962,6 +990,35 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
       BOOST_CHECK_EQUAL(to.serialize(), expected.serialize());
    }
 
+BOOST_AUTO_TEST_CASE(MergeAndRemoveSource)
+   {
+      shared::CDataContainer to(
+         "{"
+         "   \"int\": 23,"
+         "   \"string\": \"\""
+         "}");
+
+      shared::CDataContainer from(
+         "{"
+         "   \"int\": 135,"
+         "   \"string\": \"1234\""
+         "}");
+
+      const shared::CDataContainer expected(
+         "{"
+         "   \"int\": 135,"
+         "   \"string\": \"1234\""
+         "}");
+      
+
+      to.mergeFrom(from);
+
+      // Override from memory
+      from = std::string("{ \"anotherdata\" : \"skdjflsjdfhgkdjhgxjkmflsdjfglsdjfmlklsmjdf\"}");
+
+      BOOST_CHECK_EQUAL(to.serialize(), expected.serialize());
+   }
+
    BOOST_AUTO_TEST_CASE(CheckDataLimits)
    {
 	   shared::CDataContainer c(
@@ -978,15 +1035,6 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
 		   "   \"int64_min\": -9223372036854775808,"
 		   "   \"int64_max\": 9223372036854775807,"
 		   "   \"uint64_max\": 18446744073709551615"
-		   "}");
-
-	   shared::CDataContainer c2(
-		   "{"
-   	       "   \"double\": 42.0,"
-		   "   \"float\": 42.0,"
-		   "   \"float_gt_ushort_max\": 65535.1,"
-		   "   \"float_gt_short_max\": 32767.1,"
-		   "   \"float_eq_short_max\": 32767.0"
 		   "}");
 
 	   //check some out of range
@@ -1357,17 +1405,17 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
    }
 
 
-#define BOOST_CHECK_MAPS(input, output) \
-   { \
-      BOOST_CHECK_EQUAL(input.size(), output.size());\
-      std::map<std::string, std::string>::iterator ii, io;\
-      io = output.begin();\
-      for (ii = input.begin(); ii != input.end(); ++ii)\
-      {\
-         BOOST_CHECK_EQUAL(ii->first, io->first);\
-         BOOST_CHECK_EQUAL(ii->second, io->second);\
-         ++io;\
-      }\
+   template <typename T>
+   void CHECK_MAPS(const std::map<std::string, T>& input, const std::map<std::string, T>& output)
+   {
+      BOOST_CHECK_EQUAL(input.size(), output.size());
+      std::map<std::string, T>::const_iterator io = output.begin();
+      for (const auto& ii : input)
+      {
+         BOOST_CHECK_EQUAL(ii.first, io->first);
+         BOOST_CHECK_EQUAL(ii.second, io->second);
+         ++io;
+      }
    }
 
    BOOST_AUTO_TEST_CASE(Map)
@@ -1375,13 +1423,56 @@ BOOST_AUTO_TEST_SUITE(TestDataContainer)
       std::map<std::string, std::string> input = {{"key1", "value1"},{"key2", "value2"},{"key3", "value3"},{"key4", "value4"}};
       shared::CDataContainer dc(input);
 
-      auto output = dc.getAsMap();
+      auto output = dc.getAsMap<std::string>();
 
-      //dont use BOOST_CHECK_EQUAL_COLLECTIONS because it do not builds with std::map
-      BOOST_CHECK_MAPS(input, output);
+      //don't use BOOST_CHECK_EQUAL_COLLECTIONS because it do not builds with std::map
+      CHECK_MAPS<std::string>(input, output);
 
       auto output2 = dc.get<std::map<std::string, std::string>>();
-      BOOST_CHECK_MAPS(input, output2);
+      CHECK_MAPS<std::string>(input, output2);
+   }
+
+   BOOST_AUTO_TEST_CASE(MapOfContainers)
+   {
+      const shared::CDataContainer input("\
+      {\
+         \"0\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0\"\
+         },\
+         \"1\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.2\"\
+         },\
+         \"2\": {\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0-rc.1\"\
+         }\
+      }");
+
+      const std::map<std::string, shared::CDataContainer> expected =
+      {
+         {"0", shared::CDataContainer("{\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0\"\
+         }")},
+         {"1", shared::CDataContainer("{\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.2\"\
+         }")},
+         {"2", shared::CDataContainer("{\
+            \"name\": \"moon\",\
+            \"description\": null,\
+            \"version\": \"1.0.0-rc.1\"\
+         }")}
+      };
+      
+      CHECK_MAPS<shared::CDataContainer>(input.getAsMap<shared::CDataContainer>(), expected);
    }
 
 BOOST_AUTO_TEST_SUITE_END()